+++ rust/2019/problem5/src/main.rs
@@ -12,7 +12,7 @@ fn main() {
     println!("The solution to part 2 is {}", part2);
 }
 
-fn solve(s: &str) -> i32 {
+fn solve(s: &str) -> i64 {
     let parsed_input = parse_program(s);
     let mut m = Machine::new_test_mode(&parsed_input, &vec![1]);
     let state = m.run();
@@ -22,7 +22,7 @@ fn solve(s: &str) -> i32 {
     }
 }
 
-fn solve2(s: &str) -> i32 {
+fn solve2(s: &str) -> i64 {
     let parsed_input = parse_program(s);
     let mut m = Machine::new_feedback_mode(&parsed_input);
     m.input(5);
